chore: untangle common and telemetry server deps
