+++ genkit-tools/package.json
@@ -9,10 +9,11 @@
     "build:common": "cd common && pnpm build && cd .. && pnpm export:schemas",
     "build:plugins": "pnpm -r --workspace-concurrency 8 -F \"./plugins/**\" build",
     "export:schemas": "npx tsx scripts/schema-exporter.ts .",
-    "pack:all": "pnpm run pack:cli && pnpm run pack:common && pnpm run pack:plugins",
+    "pack:all": "pnpm run pack:cli && pnpm run pack:telemetry-server && pnpm run pack:common && pnpm run pack:plugins",
     "pack:common": "cd common && pnpm pack --pack-destination ../../dist",
     "pack:cli": "cd cli && pnpm pack --pack-destination ../../dist",
-    "pack:plugins": "for i in plugins/*/; do cd $i && pnpm pack --pack-destination ../../../dist && cd ../..; done"
+    "pack:plugins": "for i in plugins/*/; do cd $i && pnpm pack --pack-destination ../../../dist && cd ../..; done",
+    "pack:telemetry-server": "cd telemetry-server && pnpm pack --pack-destination ../../dist"
   },
   "devDependencies": {
     "json-schema": "^0.4.0",
