+++ genkit-tools/telemetry-server/package.json
@@ -7,7 +7,7 @@
     "genai",
     "generative-ai"
   ],
-  "version": "0.5.11",
+  "version": "0.6.0-dev.1",
   "type": "commonjs",
   "scripts": {
     "compile": "tsc -b ./tsconfig.cjs.json ./tsconfig.esm.json ./tsconfig.types.json",
@@ -24,6 +24,7 @@
   "author": "genkit",
   "license": "Apache-2.0",
   "dependencies": {
+    "@genkit-ai/tools-common": "workspace:*",
     "@asteasolutions/zod-to-openapi": "^7.0.0",
     "@google-cloud/firestore": "^7.6.0",
     "@opentelemetry/api": "^1.9.0",
